An rdp dockerfile (#3)
A linux container with xrdp installed to allow building the game in a linux environment, while also being able to interact with the game on a screen. Rather than require linux, using a docker environment makes it more portable.

The VNC solution was having significant keyboard delay when performing actions. This made it difficult to diagnose some of the errors I was seeing with movement and collision.