Pull in XGameLib into the project (#4)
Bring the XGameLib project into the monorepo. This does not import any of the repository history, as I will be preserving the original repository. As for per-file history, the project has been renamed/reorganized before, so I do not believe the history will be especially useful for diagnosing changes.

I will not bring in the bazel build process just yet, and instead inline the xgamelib for now.