Integrate XSamples into the project, built with Bazel (#6)
Integrate in the samples available from the xsamples repository, ensuring that they can all be built with bazel. This simplifies the building and compilation targets of the files, as we no longer need to worry about per-directory makefiles.

As with XGameLib, the history of these files has not been preserved, and can be referenced from the original repositories.

The samples in this project are all built with Bazel.